+++ Strategy/StrategyMultiplier.cpp
@@ -5,7 +5,7 @@ StrategyMultiplier::StrategyMultiplier()
     
 }
 
-unsigned int StrategyMultiplier::getBet(unsigned int trueCount) const
+unsigned int StrategyMultiplier::getBet(double trueCount) const
 {
     return 2;
 }
